Optimization for shaders
1. Removed dynamic branch for PCSS, PCF, Default with Poisson Sample.
  -> Dynamic Branches with PCSS, PCF, Default made significant performance down. funny things it is not affected simple branches like Collided or UseAmbientLight.
2. Added the function shader could enable functions by using preprocessor that is set before compiling shader time.
3. All precision highp -> mediump

+++ WebGL/06_Shadow/shaders/shadowmap/fs_cubemap.glsl
@@ -2,8 +2,8 @@
 
 #include "common.glsl"
 
-precision highp float;
-precision highp sampler2DArray;
+precision mediump float;
+precision mediump sampler2DArray;
 
 in vec3 Normal_;
